// types/types.ts
export type TabKeys =
  | "home"
  | "ads"
  | "messages"
  | "billing"
  | "support"
  | "supportQuestions"
  | "supportTicket"
  | "plugins"
  | "projects"
  | "createpostform"
  | "karjooform"
  | "karfarmaform"
  | "adsform"
  | "digitalprojectform"
  | "myads"
  | "chat"
  | "adsdetails"
  | "chatdetails"
  | "adminOptions"
  | "projectdetails";